-added in hud system -changed adding drawable objects so that texture is passed in -standardized scaling of objects and tiling of textures -general cleanup -menu is more generalized as far as cross screen compatibility

+++ src/org/alchemicstudio/BaseObject.java
@@ -32,7 +32,7 @@ public class BaseObject {
 	 * 
 	 * @param timeDelta
 	 */
-	public void update(float timeDelta) {
+	public void update(long timeDelta) {
 		
 	}
 }
